# api/openapi.yaml
openapi: 3.1.0
info:
  title: Domain Monitoring & Asset Intelligence API
  version: 0.6.0-phase6
  description: Bilingual Thai/English domain monitoring, Drive/Excel imports, rule-based recommendations and reporting API.
servers:
  - url: /api/v1
paths:
...
  /google-sheets/imports:
    get:
      summary: List Google Sheet import history
      security: [{sessionCookie: []}]
      responses:
        "200": {description: Import history}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: domainintel_session
...